+++ kernel/sysproc.c
@@ -9,7 +9,7 @@
 #include "sysinfo.h"
 
 uint64 freemem(void);
-uint nproc(void);
+uint64 nproc(void);
 
 uint64
 sys_exit(void)
